+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Please refer to the [official documentation](https://wilfred.readthedocs.io/en/l
 * **Changed** Removed unnecessary `python3-distutils` and `build-essential` from `stage-packages` within the Snapcraft configuration (see [this comment](https://github.com/wilfred-dev/wilfred/issues/30#issuecomment-581396779)). 53 MB decrease in package size.
 * **Fixed** [#17](https://github.com/wilfred-dev/wilfred/issues/17) Changing port using `wilfred edit` should be able to trigger configuration update on supported filetypes (this is closely related to image API v2 and [#21](https://github.com/wilfred-dev/wilfred/issues/21)).
 * **Fixed** [#28](https://github.com/wilfred-dev/wilfred/issues/28) SQLAlchemy exceptions no longer occur when trying to delete a server that has no environment variables.
+* **Fixed** [#31](https://github.com/wilfred-dev/wilfred/issues/31) Config settings that are linked to an environment variable are no longer editable using `wilfred config`.
 
 ## v0.3.0 (released on 2020-01-25)